# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  cpf: string;
  senha: string;
}

export interface LoginResult {
  token: string;
  id: number;
  nome: string;
  cpf: string;
  quarto: number;
}

export interface Hospede {
  id: number;
  nome: string;
  cpf: string;
  quarto: number;
  telefone: string;
}

export interface HospedeInput {
  nome: string;
  cpf: string;
  quarto: number;
  telefone: string;
  senha: string;
}

export type DeleteResultDeleted = {
  id?: number;
};

export interface DeleteResult {
  ok: boolean;
  deleted?: DeleteResultDeleted;
}