# lib/api-zod/src/generated/types/deleteResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DeleteResultDeleted } from './deleteResultDeleted';

export interface DeleteResult {
  ok: boolean;
  deleted?: DeleteResultDeleted;
}