# lib/api-zod/src/generated/types/hospede.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Hospede {
  id: number;
  nome: string;
  cpf: string;
  quarto: number;
  telefone: string;
}